refactor(agent): extract request-client mixins from run_agent.py (shard s4)
