+++ plugins/agent/__init__.py
@@ -1,0 +1,1 @@
+# Agent mixin package — extracted modules from run_agent.py (god-file decomposition wave 1).
