+++ plugins/agent/mixins/__init__.py
@@ -1,0 +1,1 @@
+# Mixins extracted from run_agent.py (god-file decomposition wave 1).
